refactor(web): tighten blog date helpers
